requirements.txt
Flask==2.0.1
PyMySQL==1.0.2
Flask-Session==0.4
black==21.9b0
Jinja2==3.0
SQLAlchemy==1.3.10
Flask-SQLAlchemy==2.4.1
requests==2.22.0
Flask-OAuth==0.12
Authlib==0.15.5
python-dotenv==0.19.2
oauthlib==3.1.1
Flask-Login==0.5.0
pyOpenSSL==19.0.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
cryptography==36.0.0
mysql-connector-python==8.0.27
Flask-Cors==3.0.10
pycodestyle==2.8.0